Simplifying constexpr example
